requirements.txt
aiohttp==3.8.1
blinker==1.4
boto==2.49.0
bpython==0.22.1
brotli==1.0.9
brotlicffi==1.0.9.2
ca_certs_locater==1.0
chardet==4.0.0
click==7.1.2
colorama==0.4.4
crcmod==1.7
cx_Oracle==8.3.0
Cython==0.29.30
django_pdb==0.6.2
dl==0.1.0
docutils==0.18.1
grpc_reflection==1.0.0
grpc_tools==1.0.0
httpretty==1.1.4
importlib_metadata==4.0.1
ipaddr==2.2.0
ipdb==0.13.9
ipython==8.4.0
jose==1.0.0
jupyterlab==3.4.2
keyring==23.0.1
libcst==0.4.3
lockfile==0.12.2
Markdown==3.3.7
mock==4.0.3
mongoengine==0.24.1
mypy_extensions==0.4.3
MySQL-python==1.2.5
notebook==6.4.11
numpy==1.19.5
oauth2client==4.1.3
onelogin==2.0.3
ordereddict==1.1
Pillow==9.1.1
prompt_toolkit==3.0.29
psycopg2==2.9.3
psycopg2cffi==2.9.0
ptpython==3.0.20
pycryptodome==3.14.1
pycurl==7.45.1
pydot==1.4.2
pydotplus==2.0.2
Pygments==2.7.1
pygraphviz==1.9
pylibmc==1.6.1
pymemcache==3.5.2
pyOpenSSL==22.0.0
python_bcrypt==0.3.2
python_memcached==1.59
pyu2f==0.1.5
pywatchman==1.4.1
PyYAML==6.0
railroad==0.5.0
sets==0.3.2
shortuuid==1.0.9
simplejson==3.17.6
tblib==1.7.0
tensorflow==2.9.1
typing_extensions==4.2.0
unicodedata2==14.0.0
vobject==0.9.6.1
Werkzeug==1.0.1
wincertstore==0.2.1
xmlrpclib==1.0.1
